Added example to send data using API Key
Created two different folders with examples for sending data using API Key and API Token
