+++ using_API_key/requirements.txt
@@ -1,0 +1,5 @@
+certifi==2022.12.7
+charset-normalizer==3.1.0
+idna==3.4
+requests==2.29.0
+urllib3==1.26.15
