+++ using_API_token/configuration.json
@@ -1,0 +1,16 @@
+{
+  "FW_CRED": {
+    "USERNAME": "fogwing_iothub_username",
+    "PASSWORD": "fogwing_iothub_username"
+  },
+  "PAYLOAD": {
+    "Temperature": 23.59,
+    "DO": 89.92,
+    "ORP": 480,
+    "Turbidity": 152,
+    "pH": 7.02,
+    "Conductivity": 1300,
+    "TDS": 452,
+    "battery_level": 89
+  }
+}
